requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
python-dotenv==1.0.0
supabase==2.4.2
httpx==0.27.0
psycopg[binary]
